moved encoding to utils, since there's nothing much left to justify a separate package
