+++ utils/codec_test.go
@@ -1,4 +1,4 @@
-package encoding
+package utils
 
 import (
 	"testing"
